+++ tests/Delay/FixedDelayLineFilter.cpp
@@ -24,35 +24,35 @@
 
 BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_line100_delay_1_test )
 {
-  ATK::FixedDelayLineFilter<float> filter(100);
+  ATK::FixedDelayLineFilter<double> filter(100);
   BOOST_CHECK_THROW(filter.set_delay(0), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_line100_delay_100_test )
 {
-  ATK::FixedDelayLineFilter<float> filter(100);
+  ATK::FixedDelayLineFilter<double> filter(100);
   BOOST_CHECK_THROW(filter.set_delay(100), std::out_of_range);
 }
 
 BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_sinus_line100_delay50_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
-    data[i] = std::sin(2 * boost::math::constants::pi<float>() * (i+1.)/48000 * 1000);
+    data[i] = std::sin(2 * boost::math::constants::pi<double>() * (i+1.)/48000 * 1000);
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::FixedDelayLineFilter<float> filter(100);
+  ATK::FixedDelayLineFilter<double> filter(100);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_delay(50);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
 
@@ -74,22 +74,22 @@ BOOST_AUTO_TEST_CASE( FixedDelayLineFilter_sinus_line100_delay50_test )
 
 BOOST_AUTO_TEST_CASE(FixedDelayLineFilter_sinus_line100_delay50_progressive_test)
 {
-  ATK::SimpleSinusGeneratorFilter<float> generator;
+  ATK::SimpleSinusGeneratorFilter<double> generator;
   generator.set_output_sampling_rate(48000);
   generator.set_frequency(480);
 
-  ATK::FixedDelayLineFilter<float> filter(100);
+  ATK::FixedDelayLineFilter<double> filter(100);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_delay(50);
   filter.process(50);
 
-  ATK::SumFilter<float> sum;
+  ATK::SumFilter<double> sum;
   sum.set_input_sampling_rate(48000);
   sum.set_input_port(0, &filter, 0);
   sum.set_input_port(1, &generator, 0);
 
-  ATK::TriangleCheckerFilter<float> output;
+  ATK::TriangleCheckerFilter<double> output;
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &sum, 0);
   output.set_amplitude(0);
@@ -102,22 +102,22 @@ BOOST_AUTO_TEST_CASE(FixedDelayLineFilter_sinus_line100_delay50_progressive_test
 
 BOOST_AUTO_TEST_CASE(FixedDelayLineFilter_sinus_line1000_delay50_progressive_test)
 {
-  ATK::SimpleSinusGeneratorFilter<float> generator;
+  ATK::SimpleSinusGeneratorFilter<double> generator;
   generator.set_output_sampling_rate(48000);
   generator.set_frequency(480);
 
-  ATK::FixedDelayLineFilter<float> filter(1000);
+  ATK::FixedDelayLineFilter<double> filter(1000);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_delay(50);
   filter.process(50);
 
-  ATK::SumFilter<float> sum;
+  ATK::SumFilter<double> sum;
   sum.set_input_sampling_rate(48000);
   sum.set_input_port(0, &filter, 0);
   sum.set_input_port(1, &generator, 0);
 
-  ATK::TriangleCheckerFilter<float> output;
+  ATK::TriangleCheckerFilter<double> output;
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &sum, 0);
   output.set_amplitude(0);
@@ -131,7 +131,7 @@ BOOST_AUTO_TEST_CASE(FixedDelayLineFilter_sinus_line1000_delay50_progressive_tes
 
 BOOST_AUTO_TEST_CASE(FixedDelayLineFilter_sinus_delay_test )
 {
-  ATK::FixedDelayLineFilter<float> filter(128);
+  ATK::FixedDelayLineFilter<double> filter(128);
   filter.set_delay(10);
   BOOST_CHECK_EQUAL(filter.get_delay(), 10);
 }
